pyproject.toml
[tool.poetry]
name = "eatlocal"
version = "0.7.3"
description = "eatlocal helps the user solve PyBite code challenges on their local machine"
authors = ["Russell Helmstedter <rhelmstedter@gmail.com>"]
license = "LICENSE"
readme = "README.md"
homepage = "https://github.com/rhelmstedter/eatlocal"
keywords = ["cli", "pybites"]

[tool.poetry.scripts]
eatlocal = "eatlocal.__main__:cli"

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^4.1.0"
typer = "^0.4.0"
python-dotenv = "^0.19.2"
rich = "^11.1.0"
beautifulsoup4 = "^4.10.0"
GitPython = "^3.1.26"
getkey = "^0.6.5"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[tool.pytest.ini_options]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
